added back notifications and duplicate file check
