+++ androidClient/app/src/main/AndroidManifest.xml
@@ -53,7 +53,6 @@
             android:enabled="true">
 
         </service>
-        <service android:name=".coursematerial.UploadService" />
 
         <activity
             android:name=".HomeActivity"
