+++ androidClient/app/src/main/java/com/macbitsgoa/comrades/coursematerial/Uploader.java
@@ -1,10 +1,15 @@
 package com.macbitsgoa.comrades.coursematerial;
 
+import android.app.NotificationManager;
+import android.content.Context;
+import android.media.RingtoneManager;
 import android.util.Log;
 
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DatabaseReference;
+import com.macbitsgoa.comrades.R;
+import com.macbitsgoa.comrades.persistance.Database;
 import com.squareup.okhttp.MediaType;
 import com.squareup.okhttp.OkHttpClient;
 import com.squareup.okhttp.Request;
@@ -16,23 +21,34 @@ import org.json.JSONObject;
 
 import java.io.File;
 import java.io.IOException;
+import java.security.SecureRandom;
 import java.util.concurrent.TimeUnit;
 
 import androidx.annotation.NonNull;
+import androidx.core.app.NotificationCompat;
+import androidx.work.Constraints;
 import androidx.work.Data;
+import androidx.work.NetworkType;
+import androidx.work.OneTimeWorkRequest;
+import androidx.work.WorkManager;
 import androidx.work.Worker;
 
 import static com.macbitsgoa.comrades.CHCKt.TAG_PREFIX;
 import static com.macbitsgoa.comrades.CHCKt.getCourseMaterialRef;
-import static com.macbitsgoa.comrades.coursematerial.UploadService.AUTHORIZATION_FIELD_KEY;
-import static com.macbitsgoa.comrades.coursematerial.UploadService.AUTHORIZATION_FIELD_VALUE_PREFIX;
-import static com.macbitsgoa.comrades.coursematerial.UploadService.DRIVE_API_BASE_URL;
-import static com.macbitsgoa.comrades.coursematerial.UploadService.fileToBytes;
-import static com.macbitsgoa.comrades.coursematerial.UploadService.getFileExtension;
-import static com.macbitsgoa.comrades.coursematerial.UploadService.getMimeType;
+import static com.macbitsgoa.comrades.coursematerial.UploadUtil.AUTHORIZATION_FIELD_KEY;
+import static com.macbitsgoa.comrades.coursematerial.UploadUtil.AUTHORIZATION_FIELD_VALUE_PREFIX;
+import static com.macbitsgoa.comrades.coursematerial.UploadUtil.DRIVE_API_BASE_URL;
+import static com.macbitsgoa.comrades.coursematerial.UploadUtil.calculateMD5;
+import static com.macbitsgoa.comrades.coursematerial.UploadUtil.fileToBytes;
+import static com.macbitsgoa.comrades.coursematerial.UploadUtil.getFileExtension;
+import static com.macbitsgoa.comrades.coursematerial.UploadUtil.getMimeType;
+import static com.macbitsgoa.comrades.coursematerial.UploadUtil.sendNotification;
 
 /**
  * {@link Worker} for uploading files.
+ * Use {@link #upload(String, String, String, String)} to easily upload the file.
+ *
+ * @author Aayush Singla.
  * @author Rushikesh Jogdand.
  */
 public class Uploader extends Worker {
@@ -48,6 +64,35 @@ public class Uploader extends Worker {
     private String fileId;
     private long fileSize;
     private String fileHash;
+    private NotificationManager notificationManager;
+    private NotificationCompat.Builder builder;
+    private int notificationId;
+
+    /**
+     * Handy method to access this class.
+     *
+     * @param filePath    of the file to be uploaded.
+     * @param accessToken of user's google drive.
+     * @param fileName    by which this file will be saved in db.
+     * @param courseId    to which this file belongs.
+     */
+    public static void upload(final String filePath, final String accessToken, final String fileName, final String courseId) {
+        Data uploaderData = new Data.Builder()
+                .putString(KEY_PATH, filePath)
+                .putString(KEY_ACCESS_TOKEN, accessToken)
+                .putString(KEY_FILE_NAME, fileName)
+                .putString(KEY_COURSE_ID, courseId)
+                .build();
+        Constraints uploadWorkConstraints = new Constraints.Builder()
+                .setRequiredNetworkType(NetworkType.CONNECTED)
+                .build();
+        OneTimeWorkRequest uploadRequest = new OneTimeWorkRequest.Builder(Uploader.class)
+                .setConstraints(uploadWorkConstraints)
+                .setInputData(uploaderData)
+                .build();
+        WorkManager.getInstance()
+                .enqueue(uploadRequest);
+    }
 
     @NonNull
     @Override
@@ -58,33 +103,59 @@ public class Uploader extends Worker {
         fileName = input.getString(KEY_FILE_NAME);
         courseId = input.getString(KEY_COURSE_ID);
 
-        fileHash = UploadService.calculateMD5(new File(path));
+        notificationId = (new SecureRandom()).nextInt(100);
+        notificationManager = (NotificationManager)
+                getApplicationContext().getSystemService(Context.NOTIFICATION_SERVICE);
+
+        notifyInit();
+        fileHash = calculateMD5(new File(path));
+        CourseMaterial duplicate = findDuplicate();
+        if (duplicate != null) {
+            notifyDuplicate(duplicate.getFileName());
+            return Result.FAILURE;
+        }
+
         String response = uploadFile();
-        if (response == null) return Result.FAILURE;
+        if (response == null) {
+            notifyFailure();
+            return Result.FAILURE;
+        }
         final JSONObject jsonObject;
         try {
             jsonObject = new JSONObject(response);
             fileId = (String) jsonObject.get("id");
         } catch (JSONException e) {
             Log.e(TAG, e.getMessage(), e.fillInStackTrace());
+            notifyFailure();
             return Result.FAILURE;
         }
         if (!makePublic(accessToken, fileId)) {
+            notifyFailure();
             return Result.FAILURE;
         }
         JSONObject metadata = obtainMetadata(fileId, accessToken);
         if (metadata == null) {
+            notifyFailure();
             return Result.FAILURE;
         }
         try {
             writeToDb(metadata);
         } catch (JSONException e) {
             Log.e(TAG, e.getMessage(), e.fillInStackTrace());
+            notifyFailure();
             return Result.FAILURE;
         }
+        notifySuccess();
         return Result.SUCCESS;
     }
 
+    /**
+     * Get metadata for the file.
+     *
+     * @param fileId      of query file.
+     * @param accessToken of owner.
+     * @return {@link JSONObject} result.
+     */
     public static JSONObject obtainMetadata(String fileId, String accessToken) {
         final OkHttpClient okHttpClient = new OkHttpClient();
         final Request request = new Request.Builder()
@@ -135,6 +206,13 @@ public class Uploader extends Worker {
         return null;
     }
 
+    /**
+     * Make the file with fileId public.
+     *
+     * @param accessToken of owner.
+     * @param fileId      of corresponding file.
+     * @return whether making the file succeeded.
+     */
     public static boolean makePublic(String accessToken, String fileId) {
         final JSONObject jsonPermission;
         try {
@@ -195,5 +273,47 @@ public class Uploader extends Worker {
         itemCourseMaterial.setLink(jsonObject.get("webContentLink").toString());
         itemCourseMaterial.setMimeType(jsonObject.get("mimeType").toString());
         materialNode.child(fileHash).setValue(itemCourseMaterial);
+    }
+
+    /**
+     * Check if the file is duplicate
+     *
+     * @return null if no duplicate exists, otherwise original file.
+     */
+    private CourseMaterial findDuplicate() {
+        return Database.getInstance(getApplicationContext()).getMaterialDao().checkHashId(fileHash);
+    }
+
+    private void notifyDuplicate(String originalFileName) {
+        builder = sendNotification(getApplicationContext(), R.drawable.ic_cloud_done_black_24dp, fileName + " couldn't be uploaded",
+                "A similar file already exists in this course with name " + originalFileName);
+        builder.setProgress(0, 0, false);
+        builder.setOngoing(false);
+        builder.setSound(RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION));
+        notificationManager.notify(notificationId, builder.build());
+    }
+
+    private void notifyInit() {
+        builder = sendNotification(getApplicationContext(), R.drawable.ic_launcher_foreground, "Uploading " + fileName, "Please wait...");
+        builder.setProgress(0, 0, true);
+        builder.setOngoing(true);
+        notificationManager.notify(notificationId, builder.build());
+    }
+
+    private void notifyFailure() {
+        builder = sendNotification(getApplicationContext(), R.drawable.ic_launcher_foreground, "Comrades",
+                "File Could not be uploaded.Please try again later.");
+        builder.setProgress(0, 0, false);
+        builder.setOngoing(false);
+        notificationManager.notify(notificationId, builder.build());
+    }
+
+    private void notifySuccess() {
+        builder = sendNotification(getApplicationContext(), R.drawable.ic_cloud_done_black_24dp, fileName + " uploaded",
+                "Thanks for Contributing");
+        builder.setProgress(0, 0, false);
+        builder.setOngoing(false);
+        builder.setSound(RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION));
+        notificationManager.notify(notificationId, builder.build());
     }
 }
